#159: use deleteReply to consolidate viewblackbox flow
